Cambia tipo de variable de numeroDescargas de double a integer en record DatosLibros, cambia tipo de variable de fechaDeNacimiento y fechaDeDeceso de String a Integer en record DatosAutor, inicia métodos para buscar autores y para mostrar autores vivos en determinado tiempo en clase Principal

+++ src/main/java/com/aluracursos/literalura/model/DatosLibros.java
@@ -10,6 +10,6 @@ public record DatosLibros(
     @JsonAlias("title") String titulo,
     @JsonAlias("authors") List<DatosAutor> autor,
     @JsonAlias("languages") List<String> idioma,
-    @JsonAlias("download_count") Double numeroDescargas
+    @JsonAlias("download_count") Integer numeroDescargas
 ) {
 }
